Moved the listMetadataObjects utility from fetch.ts to utils.ts
