+++ packages/salesforce-adapter/src/adapter.ts
@@ -99,8 +99,8 @@ import changedAtSingletonFilter from './filters/changed_at_singleton'
 import { FetchElements, FetchProfile, SalesforceConfig } from './types'
 import { getConfigFromConfigChanges } from './config_change'
 import { LocalFilterCreator, Filter, FilterResult, RemoteFilterCreator, LocalFilterCreatorDefinition, RemoteFilterCreatorDefinition } from './filter'
-import { addDefaults, isCustomObjectSync, isCustomType } from './filters/utils'
-import { retrieveMetadataInstances, fetchMetadataType, fetchMetadataInstances, listMetadataObjects } from './fetch'
+import { addDefaults, isCustomObjectSync, isCustomType, listMetadataObjects } from './filters/utils'
+import { retrieveMetadataInstances, fetchMetadataType, fetchMetadataInstances } from './fetch'
 import { isCustomObjectInstanceChanges, deployCustomObjectInstancesGroup } from './custom_object_instances_deploy'
 import { getLookUpName, getLookupNameWithFallbackToElement } from './transformers/reference_mapping'
 import { deployMetadata, NestedMetadataTypeInfo } from './metadata_deploy'
