+++ packages/salesforce-adapter/src/fetch.ts
@@ -17,28 +17,29 @@ import _ from 'lodash'
 import { safeJsonStringify } from '@salto-io/adapter-utils'
 import { FileProperties, MetadataInfo, MetadataObject } from 'jsforce-types'
 import { InstanceElement, ObjectType, TypeElement } from '@salto-io/adapter-api'
-import { values as lowerDashValues, collections } from '@salto-io/lowerdash'
+import { collections, values as lowerDashValues } from '@salto-io/lowerdash'
 import { logger } from '@salto-io/logging'
 import {
-  FetchElements,
   ConfigChangeSuggestion,
-  MAX_ITEMS_IN_RETRIEVE_REQUEST,
+  FetchElements,
+  FetchProfile,
   MAX_INSTANCES_PER_TYPE,
-  MetadataQuery, FetchProfile,
+  MAX_ITEMS_IN_RETRIEVE_REQUEST,
+  MetadataQuery,
 } from './types'
 import {
-  METADATA_CONTENT_FIELD,
-  NAMESPACE_SEPARATOR,
-  INTERNAL_ID_FIELD,
-  DEFAULT_NAMESPACE,
-  RETRIEVE_SIZE_LIMIT_ERROR,
-  LAYOUT_TYPE_ID_METADATA_TYPE,
+  API_NAME_SEPARATOR,
   CUSTOM_OBJECT,
-  UNLIMITED_INSTANCES_VALUE,
+  DEFAULT_NAMESPACE,
   FOLDER_CONTENT_TYPE,
   INSTALLED_PACKAGE_METADATA,
-  API_NAME_SEPARATOR,
+  INTERNAL_ID_FIELD,
+  LAYOUT_TYPE_ID_METADATA_TYPE,
+  METADATA_CONTENT_FIELD,
+  NAMESPACE_SEPARATOR,
   PROFILE_METADATA_TYPE,
+  RETRIEVE_SIZE_LIMIT_ERROR,
+  UNLIMITED_INSTANCES_VALUE,
 } from './constants'
 import SalesforceClient, { ErrorFilter } from './client/client'
 import {
@@ -47,11 +48,18 @@ import {
   createSkippedListConfigChange,
   createSkippedListConfigChangeFromError,
 } from './config_change'
-import { apiName, createInstanceElement, MetadataObjectType, createMetadataTypeElements, getAuthorAnnotations } from './transformers/transformer'
-import { fromRetrieveResult, toRetrieveRequest, getManifestTypeName } from './transformers/xml_transformer'
+import {
+  apiName,
+  createInstanceElement,
+  createMetadataTypeElements,
+  getAuthorAnnotations,
+  MetadataObjectType,
+} from './transformers/transformer'
+import { fromRetrieveResult, getManifestTypeName, toRetrieveRequest } from './transformers/xml_transformer'
+import { listMetadataObjects } from './filters/utils'
 
 const { isDefined } = lowerDashValues
-const { makeArray, splitDuplicates } = collections.array
+const { makeArray } = collections.array
 const { awu, keyByAsync } = collections.asynciterable
 const log = logger(module)
 
@@ -107,35 +115,6 @@ const withFullPath = (props: FileProperties, folderPathByName: Record<string, st
       fileName: props.fileName.replace(folderName, fullPath),
     }
     : props
-}
-
-const removeDuplicateFileProps = (files: FileProperties[]): FileProperties[] => {
-  const { duplicates, uniques } = splitDuplicates(files, fileProps => `${fileProps.namespacePrefix}__${fileProps.fullName}`)
-  duplicates.forEach(props => {
-    log.warn('Found duplicate file props with the same name in response to listMetadataObjects: %o', props)
-  })
-  return uniques.concat(duplicates.map(props => props[0]))
-}
-
-export const listMetadataObjects = async (
-  client: SalesforceClient,
-  metadataTypeName: string,
-  isUnhandledError?: ErrorFilter,
-): Promise<FetchElements<FileProperties[]>> => {
-  const { result, errors } = await client.listMetadataObjects(
-    { type: metadataTypeName },
-    isUnhandledError,
-  )
-
-  // Salesforce quirk, we sometimes get the same metadata fullName more than once
-  const elements = removeDuplicateFileProps(result)
-
-  return {
-    elements,
-    configChanges: errors
-      .map(e => e.input)
-      .map(createListMetadataObjectsConfigChange),
-  }
 }
 
 const getNamespace = (obj: FileProperties): string => (
